requirements.txt
absl-py==2.1.0
annotated-types==0.5.0
anyio==3.7.1
astor==0.8.1
boto3==1.33.13
botocore==1.33.13
cached-property==1.5.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1707022139797/work/certifi
click==8.1.7
colorama==0.4.6
distro==1.9.0
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.2.0
fastapi==0.103.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.60.1
h11==0.14.0
h5py==3.8.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.6
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
llvmlite==0.39.1
Markdown==3.4.4
MarkupSafe==2.1.5
numba==0.56.4
numpy==1.21.6
openai==1.12.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
orjson==3.9.7
pathlib==1.0.1
Pillow==9.5.0
protobuf==3.20.0
pydantic==2.5.3
pydantic-extra-types==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.14.6
pydub==0.25.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.8
python-speech-features==0.6
PyYAML==6.0.1
resampy==0.2.2
s3transfer==0.8.2
scipy==1.5.4
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tensorboard==1.15.0
tensorflow==1.15.2
tensorflow-estimator==1.15.1
termcolor==2.3.0
torch==1.7.1+cu110
torchvision==0.8.2+cu110
tqdm==4.66.2
typing_extensions==4.7.1
ujson==5.7.0
urllib3==1.26.18
uvicorn==0.22.0
watchfiles==0.20.0
websockets==11.0.3
Werkzeug==2.2.3
wincertstore==0.2
wrapt==1.16.0
zipp==3.15.0
